contribution/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Flask-Mail==0.9.1
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
paystackapi==2.1.0
Flask-Migrate==4.0.5
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
email-validator==2.1.0
python-dateutil==2.8.2
